+++ src/main/java/ru/fev/accumulation/service/ClientServiceImpl.java
@@ -1,12 +1,18 @@
 package ru.fev.accumulation.service;
 
+import jakarta.persistence.PersistenceException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import ru.fev.accumulation.entity.Client;
+import ru.fev.accumulation.exceptions.PAEntityNotFoundException;
+import ru.fev.accumulation.exceptions.PAIllegalIdException;
+import ru.fev.accumulation.exceptions.PAIncorrectArgumentException;
 import ru.fev.accumulation.repository.ClientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.security.InvalidParameterException;
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -18,8 +24,9 @@ public class ClientServiceImpl implements ClientService {
     @Override
     public void addClient(Client client) {
 
-        if (client.getCardNumber().length() != Client.CARD_NUMBER_LENGTH) {
-            throw new InvalidParameterException("Incorrect card number");
+        if (client.getCardNumber().length() != Client.CARD_NUMBER_LENGTH
+                || client.getCardNumber().isBlank()) {
+            throw new PAIncorrectArgumentException("Incorrect card number");
         }
 
         clientRepository.save(client);
@@ -27,20 +34,51 @@ public class ClientServiceImpl implements ClientService {
 
     @Override
     public Client getById(Long id) {
-        return clientRepository.getReferenceById(id);
+        if (id < 1) {
+            throw new IllegalArgumentException("Id must be greater than zero");
+        }
+
+        try {
+            return clientRepository.getReferenceById(id);
+        } catch (Exception e) {
+            throw new PAEntityNotFoundException("Incorrect ID");
+        }
     }
 
     public Client getByCardNumber(String cardNumber) {
-        return clientRepository.getByCardNumber(cardNumber);
+
+        if (cardNumber.isBlank() || cardNumber.length() != Client.CARD_NUMBER_LENGTH) {
+            throw new PAIncorrectArgumentException("Incorrect card number");
+        }
+        List<Integer> list = new ArrayList<>();
+        try {
+            return clientRepository.getByCardNumber(cardNumber);
+        } catch (PersistenceException e) {
+            throw new PAEntityNotFoundException("Client with card {cardNumber} not found");
+        }
     }
 
     @Override
     public int getDiscountPoints(Long id) {
-        return clientRepository.getReferenceById(id).getDiscountPoints();
+        if (id < 1) {
+            throw new PAIllegalIdException("Check ID must be greater than zero");
+        }
+
+        try {
+            return clientRepository
+                    .getReferenceById(id)
+                    .getDiscountPoints();
+        } catch (Exception e) {
+            throw new PAEntityNotFoundException("Incorrect ID");
+        }
     }
 
     @Override
     public void deleteClient(Long id) {
+        if (id < 1) {
+            throw new PAIllegalIdException("Check ID must be greater than zero");
+        }
+
         clientRepository.deleteById(id);
     }
 
@@ -52,6 +90,14 @@ public class ClientServiceImpl implements ClientService {
     @Transactional
     @Override
     public void subtractDiscountPoints(Long id, int pointsToSubtract) {
+        if (id < 1) {
+            throw new PAIllegalIdException("Check ID must be greater than zero");
+        }
+
+        if (this.getDiscountPoints(id) < pointsToSubtract) {
+           throw new PAIncorrectArgumentException("Incorrect value of 'pointsToSubtract'");
+        }
+
         clientRepository.subtractDiscountPoints(id, pointsToSubtract);
     }
 }
